[DP][SUPP-430] move models to individual files and improve tests
